Wed, Aug 31, 2022, 5:33 AM +00:00

+++ filter-less/helpers.c
@@ -88,22 +88,22 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
     for (int i = 0; i < height; i++) {
         for (int j = 0; j < width / 2; j++) {
 
-            int upperRed = image[i - 1][j - 1].rgbtRed + image[i - 1][j].rgbtRed + image[i - 1][j + 1].rgbtRed;
-            int middleRed = image[i][j - 1].rgbtRed + image[i][j + 1].rgbtRed;
-            int lowerRed = image[i + 1][j - 1].rgbtRed + image[i + 1][j].rgbtRed + image[i + 1][j + 1].rgbtRed;
+            float upperRed = image[i - 1][j - 1].rgbtRed + image[i - 1][j].rgbtRed + image[i - 1][j + 1].rgbtRed;
+            float middleRed = image[i][j - 1].rgbtRed + image[i][j + 1].rgbtRed;
+            float lowerRed = image[i + 1][j - 1].rgbtRed + image[i + 1][j].rgbtRed + image[i + 1][j + 1].rgbtRed;
 
             image[i][width - 1 - j].rgbtRed = leftRed;
             image[i][j].rgbtRed = rightRed;
 
-            int upperGreen = image[i - 1][j - 1].rgbtGreen + image[i - 1][j].rgbtGreen + image[i - 1][j + 1].rgbtGreen;
-            int middleGreen = image[i][j - 1].rgbtGreen + image[i][j + 1].rgbtGreen;
-            int lowerGreen = image[i + 1][j - 1].rgbtGreen + image[i + 1][j].rgbtGreen + image[i + 1][j + 1].rgbtGreen;
+            float upperGreen = image[i - 1][j - 1].rgbtGreen + image[i - 1][j].rgbtGreen + image[i - 1][j + 1].rgbtGreen;
+            float middleGreen = image[i][j - 1].rgbtGreen + image[i][j + 1].rgbtGreen;
+            float lowerGreen = image[i + 1][j - 1].rgbtGreen + image[i + 1][j].rgbtGreen + image[i + 1][j + 1].rgbtGreen;
             image[i][width - 1 - j].rgbtGreen = leftGreen;
             image[i][j].rgbtGreen = rightGreen;
 
-            int upperBlue = image[i - 1][j - 1].rgbtBlue + image[i - 1][j].rgbtBlue + image[i - 1][j + 1].rgbtBlue;
-            int middleBlue = image[i][j - 1].rgbtBlue + image[i][j + 1].rgbtBlue;
-            int lowerBlue = image[i + 1][j - 1].rgbtBlue + image[i + 1][j].rgbtBlue + image[i + 1][j + 1].rgbtBlue;
+            float upperBlue = image[i - 1][j - 1].rgbtBlue + image[i - 1][j].rgbtBlue + image[i - 1][j + 1].rgbtBlue;
+            float middleBlue = image[i][j - 1].rgbtBlue + image[i][j + 1].rgbtBlue;
+            float lowerBlue = image[i + 1][j - 1].rgbtBlue + image[i + 1][j].rgbtBlue + image[i + 1][j + 1].rgbtBlue;
             image[i][width - 1 - j].rgbtBlue = leftBlue;
             image[i][j].rgbtBlue = rightBlue;
         }
